Reorganize tests
Follow-up from #358 to make the organization make more sense.
